fix: screens names
