+++ apps/web/src/Screens/Details/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Details';
