+++ apps/web/src/Screens/PackageSearch/PackageSearch.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function PackageSearch() {
+  return <div>PackageSearch</div>;
+}
+
+export default PackageSearch;
