+++ apps/web/src/Screens/PackageSearch/index.tsx
@@ -1,7 +1,1 @@
-import React from 'react';
-
-function PackageSearch() {
-  return <div>PackageSearch</div>;
-}
-
-export default PackageSearch;
+export { default } from './PackageSearch';
