+++ apps/web/src/Screens/Uploaded/Uploaded.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Uploaded() {
+  return <div>Uploaded</div>;
+}
+
+export default Uploaded;
