+++ apps/web/src/Screens/Uploaded/index.tsx
@@ -1,7 +1,1 @@
-import React from 'react';
-
-function Uploaded() {
-  return <div>Uploaded</div>;
-}
-
-export default Uploaded;
+export { default } from './Uploaded';
